Update Assignment-1.md
Coming to terms: What I was trying to do here was to write an informative and entertaining piece on how to begin a blog on food. I think I did just that. Comments I received on this piece said it was easy to read and understand, and I think my writing voice worked well with the piece. Overall I'm happy with it.
Forwarding: Some pros about this piece is that I happened to have already started a food blog for a different class a couple of semesters ago, so I was able to speak from a place of authority. I am also not an expert, so I was able to understand where people may have confusion and write tips that helped me along.
Countering: As far as cons of this piece, I really don't like the aestetics of it. I think it's pretty boring to look at and I wouldn't be inclinded to read it if I just found it on the internet. 
Looking ahead: I could see myself writing something similar again, like some sort of instructional blog. I think my ability to be detailed would lend itself well to that.

@matthewjmiller